Dummy 버전 만들어서 테스트 중
DGameManager.cs23부터 이어서 하기

+++ Assets/Dev/Script/Manager/GameMgr.cs
@@ -14,7 +14,7 @@ public class GameMgr
     public static int State { get => state; }
     private static int state;
 
-    public static void MapData_Change(ushort mapCode)
+    public static void MapData_Change(int mapCode)
     {
         LastMap = NowMap;
         lastFieldPos = UnitMgr.MyPC.Pos;
